backend/services/voice-agent/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
httpx==0.25.1
websockets==11.0.3

# Pydantic for data validation
pydantic==2.4.2
pydantic-settings==2.0.3

# Redis for session management and queue
redis==5.0.1

# Speech recognition
openai-whisper==20231117

# LLM and agent framework
openai==1.3.5
langgraph==0.0.19

# Text-to-speech
elevenlabs==0.2.26

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1 